requirements.txt
gradio
torch
transformers
faiss-cpu
numpy
pillow
tqdm
kagglehub
